backend/database.py
import mysql.connector
import os
from dotenv import load_dotenv

# Load environment variables from the .env file
load_dotenv()

class Database:
    def __init__(self):
        self.connection = None
        self.cursor = None

    def connect(self):
        try:
            self.connection = mysql.connector.connect(
                host=os.getenv("MYSQL_HOST"),
                user=os.getenv("MYSQL_USER"),
                password=os.getenv("MYSQL_PASSWORD"),
                database=os.getenv("MYSQL_DATABASE")
            )
            self.cursor = self.connection.cursor(dictionary=True) # dictionary=True makes rows accessible by column name
            print("Successfully connected to MySQL database.")
        except mysql.connector.Error as err:
            print(f"Error connecting to database: {err}")

    def close(self):
        if self.connection and self.connection.is_connected():
            self.cursor.close()
            self.connection.close()
            print("MySQL connection is closed.")

# Initialize the database class for use in other files
db = Database()