backend/requirements.txt
Flask
google-cloud-aiplatform
mysql-connector-python
python-dotenv
Flask-Cors